backend/auth/requirements.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
PyJWT>=2.8.0
